nixos: Write the directories creation script into store
When user writes a long list of directories, the original code
generates a very long activation script, which can cause an "Argument
list too long" error.

This try to fix the issue by writing the directories creation script
into nix store as an executable instead of repeating it multiple
times.