home-manager: Add `allowOther` option to let other users read bind mounts
NixOS defaults to not letting fuse mounts be allowed to let other
users read their contents. `bindfs` wants to give other users access
and is therefore normally run with `--no-allow-other` to not throw an
error.

Giving other users, mainly `root`, access to the bind mounts is,
however, useful and works fine when

programs.fuse.userAllowOther = true;

is declared in `configuration.nix`. This adds an option to choose
whether to give other users access or not. It also prompts the user to
set the `allowOther` attribute with a link to the documentation.