nixos: Add support for custom permissions on created directories
Allow the owner and mode to be set when directories are created in
persistent storage by the `create-directories.bash` script. Very
useful for directories used to store secrets.

Also, make sure the user directories are created with reasonable
defaults, i.e. owned by the user and its group, not by `root:root`.