nixos: Don't exit with an error in the activation script
Exiting with an error in the activation script causes the whole script
to fail and system boot thus fails with a kernel panic. Since the
point of exiting with an error is mainly to make the systemd service
fail, let's do it only when the script runs as a service.

Fixes #77