nixos: topologically sort directories
to be created/chowned/chmodded using `create-directories.bash`.

The directories are sorted such that parent *source* paths (that is,
paths under persistent storage directories) are processed prior to child
source paths, and parent *destination* paths (that is, paths under the
target installation root) are processed prior to child destination
paths.