nixos: Introduce `persistentStoragePath` option
This allows the path to persistent storage to be defined through an
option in the submodule, not just by the submodule name, i.e. instead
of

```nix
environment.persistence."/persistent" = {
  files = [ ... ]
  directories = [ ... ]
};
```

you could specify

```nix
environment.persistence.main = {
  persistentStoragePath = "/persistent";
  files = [ ... ]
  directories = [ ... ]
};
```

which is good for readability and when you want to split the config
between multiple files.

Fixes #117.