Fix build with `documentation.nixos.includeAllModules = true;`
This includes the options of all modules used in the evaluation, not
just the ones from `<nixpkgs/nixos>` in the local manual.

Right now this breaks with

    error: attribute '"‹name›"' missing

because the submodule `environment.persistence` doesn't have
actual declarations when building the manual, but a dummy only to evaluate
the sub-options (including their defaults which is the cause of the
error) and generate documentation from that.

Using `defaultText` prevents that because there's no need anymore to
evaluate the `default` values of the options.